requirements.txt
absl-py==1.0.0
alembic==1.7.7
attrs==21.4.0
click==8.1.3
Flask==2.1.2
flask-expects-json==1.7.0
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.5.1
Mako==1.2.0
MarkupSafe==2.1.1
numpy==1.22.3
ortools==9.3.10497
protobuf==3.20.1
pyrsistent==0.18.1
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
zipp==3.8.0
